% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbachAlpha.R
\name{cronbachAlpha}
\alias{cronbachAlpha}
\title{\code{cronbachAlpha} Object}
\description{
\code{cronbachAlpha} Object
}
\examples{
list(
 LSAT = cronbach.alpha(LSAT, CI = TRUE, B = 500)
) |> fastmd::render2html()

}
