% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{ltm.tzh-package}
\alias{ltm.tzh}
\alias{ltm.tzh-package}
\title{ltm.tzh: Extensions for ltm}
\description{
To extend ltm package.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

}
\keyword{internal}
