% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbachAlpha.R
\name{md_.cronbachAlpha}
\alias{md_.cronbachAlpha}
\title{md_.cronbachAlpha}
\usage{
\method{md_}{cronbachAlpha}(x, xnm, ...)
}
\arguments{
\item{x}{an object of class \code{'cronbachAlpha'},
returned from function \link[ltm]{cronbach.alpha}}

\item{xnm}{..}

\item{...}{..}
}
\description{
md_.cronbachAlpha
}
\examples{
library(rmd.tzh); library(ltm); list(
 cronbachAlpha = cronbach.alpha(LSAT, CI = TRUE, B = 500)
) |> render_(file = 'cronbachAlpha')
}
\keyword{internal}
