% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbachAlpha.R
\name{rmd_.cronbachAlpha}
\alias{rmd_.cronbachAlpha}
\title{rmd_.cronbachAlpha}
\usage{
rmd_.cronbachAlpha(x, xnm, ...)
}
\arguments{
\item{x}{..}

\item{xnm}{..}

\item{...}{..}
}
\description{
rmd_.cronbachAlpha
}
