% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbachAlpha.R
\name{S3_cronbachAlpha}
\alias{S3_cronbachAlpha}
\alias{endpoint.cronbachAlpha}
\alias{estnm.cronbachAlpha}
\title{S3 methods for \code{cronbachAlpha}}
\usage{
\method{endpoint}{cronbachAlpha}(x)

\method{estnm}{cronbachAlpha}(x)
}
\arguments{
\item{x}{an object of class \code{'cronbachAlpha'},
returned from function \link[ltm]{cronbach.alpha}}
}
\description{
Additional S3 methods for \code{'cronbachAlpha'}.
}
\keyword{internal}
