% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbachAlpha.R
\name{Sprintf.cronbachAlpha}
\alias{Sprintf.cronbachAlpha}
\title{Sprintf.cronbachAlpha}
\usage{
Sprintf.cronbachAlpha(model, ...)
}
\arguments{
\item{model}{..}

\item{...}{..}
}
\description{
Sprintf.cronbachAlpha
}
