% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbachAlpha.R
\name{md_.cronbachAlpha}
\alias{md_.cronbachAlpha}
\title{\code{\link[=md_.cronbachAlpha]{md_.cronbachAlpha()}}}
\usage{
\method{md_}{cronbachAlpha}(x, xnm, ...)
}
\arguments{
\item{x}{an object of class \code{'cronbachAlpha'},
returned from function \link[ltm]{cronbach.alpha}}

\item{xnm}{..}

\item{...}{..}
}
\description{
\code{\link[=md_.cronbachAlpha]{md_.cronbachAlpha()}}
}
\examples{
list(
 cronbachAlpha = cronbach.alpha(LSAT, CI = TRUE, B = 500)
) |> rmd.tzh::render_(file = 'cronbachAlpha')
}
\keyword{internal}
