% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbachAlpha.R
\name{S3_cronbachAlpha}
\alias{S3_cronbachAlpha}
\alias{endpoint.cronbachAlpha}
\alias{.pval.cronbachAlpha}
\alias{estName.cronbachAlpha}
\alias{note_.cronbachAlpha}
\alias{Sprintf.cronbachAlpha}
\title{S3 methods for \code{cronbachAlpha}}
\usage{
endpoint.cronbachAlpha(x)

.pval.cronbachAlpha(x)

estName.cronbachAlpha(x)

note_.cronbachAlpha(x)

Sprintf.cronbachAlpha(x)
}
\arguments{
\item{x}{an object of class \code{'cronbachAlpha'},
returned from function \link[ltm]{cronbach.alpha}}
}
\description{
Additional S3 methods for \code{'cronbachAlpha'} (\code{cibeta}).
}
